+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/Arbitrary.hs
@@ -11,7 +11,7 @@ instance Arbitrary MemberStatus where
   arbitrary = arbitraryBoundedEnum
 
 instance Arbitrary (QueryPoolStateResult era) where
-  arbitrary = QueryPoolStateResult <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
+  arbitrary = QueryPoolStateResult <$> arbitrary <*> arbitrary <*> arbitrary
 
 instance Arbitrary StakeSnapshot where
   arbitrary = genericArbitraryU